+++ CHANGELOG.md
@@ -9,8 +9,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Optional support for compressing with ZStandard
+- Optional support for compressing with BROTLI
 - Optional support for compressing with bzip2
+- Optional support for compressing with DEFLATE
+- Optional support for compressing with ZStandard
 
 ### Changed